refactor: add encapsulation and single responsibility principle
chore: some other cleanup/reshuffle
