+++ .env.example
@@ -3,3 +3,5 @@ GEMINI_MODEL_ID=your-gemini-model-id
 MODEL_CYPHER_ID=your-cypher-model-id
 MEMGRAPH_HOST=localhost
 MEMGRAPH_PORT=7687
+LAB_PORT=3000
+MEMGRAPH_HTTP_PORT=7444
