+++ docker-compose.yaml
@@ -3,11 +3,11 @@ services:
   memgraph:
     image: memgraph/memgraph-mage
     ports:
-      - "7687:7687"
-      - "7444:7444"
+      - "${MEMGRAPH_PORT:-7687}:7687"
+      - "${MEMGRAPH_HTTP_PORT:-7444}:7444"
   lab:
     image: memgraph/lab
     ports:
-      - "3000:3000"
+      - "${LAB_PORT:-3000}:3000"
     environment:
       QUICK_CONNECT_MG_HOST: memgraph
